refactoring initialize BoardState method
