+++ src/components/atoms/Piece/container.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Const from 'src/const';
 import Piece from 'src/components/atoms/Piece/component';
 
-import { PlayerValType } from 'src/types';
+import { PlayerValType } from 'src/@types';
 
 const { Color, PlayerVal } = Const;
 interface PieceProp {
